archie/requirements.txt
fastapi
uvicorn[standard]
tree-sitter
tree-sitter-python
tree-sitter-javascript
networkx
lancedb
sentence-transformers
openai
anthropic
PyGithub
watchdog
gitpython
python-dotenv
pydantic
pydantic-settings
pytest
pytest-asyncio
httpx
aiofiles
